+++ python/examples/4-profiling.py
@@ -29,7 +29,7 @@ from metatomic.torch import (
     ModelOutput,
     System,
 )
-from metatomic.torch.ase_calculator import MetatomicCalculator
+from metatomic_ase import MetatomicCalculator
 
 
 # %%